+++ hydra/_internal/hydra.py
@@ -28,8 +28,8 @@ from hydra.plugins.search_path_plugin import SearchPathPlugin
 from hydra.plugins.sweeper import Sweeper
 from hydra.types import RunMode, TaskFunction
 
-from ..core.callbacks import Callbacks
 from ..core.default_element import DefaultsTreeNode, InputDefault
+from .callbacks import Callbacks
 from .config_loader_impl import ConfigLoaderImpl
 from .utils import create_automatic_config_search_path
